gui-win32: fix encoding and decoding of clipboard strings
On Windows, the clipboard strings are encoded in UTF-16.
However, Drawterm considers them as Rune. It used to work
when a Rune was a short, but a Rune is now a int.
The solution is to implement Rune16 functions to
handle UTF-16 strings.

The Rune16 functions were written by Charles Forsyth
as part of Inferno. Balaji Srinivasa integrated them
into Drawterm.